% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.environment.R
\name{compare.environment}
\alias{compare.environment}
\title{Compare Environment}
\usage{
compare.environment()
}
\value{
List of environment entries that are different between the packages.
}
\description{
Compare environments in \pkg{TAF} and \pkg{icesTAF}.
}
