% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.files.R
\name{compare.files}
\alias{compare.files}
\title{Compare Files}
\usage{
compare.files(tools.prod = "~/git/ices/tools-prod")
}
\arguments{
\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}
}
\value{
List of filenames that are different between the repositories.
}
\description{
Compare \code{*.R} files in \pkg{TAF} and \pkg{icesTAF}.
}
\examples{
\dontrun{
tools.prod <- "~/git/ices/tools-prod"
compares.files(tools.prod)
}

}
