% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.namespace.R
\name{compare.namespace}
\alias{compare.namespace}
\title{Compare Namespace}
\usage{
compare.namespace(tools.prod = "~/git/ices/tools-prod")
}
\arguments{
\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}
}
\value{
List of namespace entries that are different between the repositories.
}
\description{
Compare NAMESPACE in \pkg{TAF} and \pkg{icesTAF}.
}
