% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.rdir.R
\name{compare.rdir}
\alias{compare.rdir}
\title{Compare R Directory}
\usage{
compare.rdir(tools.prod = "~/git/ices/tools-prod")
}
\arguments{
\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}
}
\value{
List of filenames that are different between the repositories.
}
\description{
Compare files in the R directory of \pkg{TAF} and \pkg{icesTAF} repositories.
}
\examples{
\dontrun{
tools.prod <- "~/git/ices/tools-prod"
compares.files(tools.prod)
}

}
