\name{state.abb.long}
\alias{state.abb.long}
\alias{state.name.long}
\title{US States (including Washington, D.C.)}
\description{
	Names and abbreviations of the 50 US States and Washington, D.C.
}
\usage{
	state.abb.long
	state.name.long
}
\details{
	This extends \link{state} by adding Washington, D.C. to state.abb and state.name.long.
}
