% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_dir.R
\name{diff_dir}
\alias{diff_dir}
\title{Compare Directory}
\usage{
diff_dir(tools.prod = "~/git/ices/tools-prod")
}
\arguments{
\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}
}
\value{
List of filenames that are different between the repositories.
}
\description{
Compare filenames in the R directory of \pkg{TAF} and \pkg{icesTAF}
repositories.
}
\examples{
\dontrun{
tools.prod <- "~/git/ices/tools-prod"
diff_dir(tools.prod)
}

}
