% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_environment.R
\name{diff_environment}
\alias{diff_environment}
\title{Compare Environment}
\usage{
diff_environment()
}
\value{
List of objects that are different between the environments.
}
\description{
Compare objects in the environments \code{package:TAF} and
\code{package:icesTAF} in an R session.
}
\examples{
diff_environment()

}
