% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_file.R
\name{diff_file}
\alias{diff_file}
\title{Compare R File}
\usage{
diff_file(file, tools.prod = "~/git/ices/tools-prod",
  ignore.internal = TRUE, ignore.noreturn = TRUE, ignore.overview = TRUE,
  simplify = TRUE)
}
\arguments{
\item{file}{name of a \file{*.R} file, without the full path.}

\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}

\item{ignore.internal}{whether to ignore lines containing the string
\code{"TAF-internal"}.}

\item{ignore.noreturn}{whether to ignore lines containing the string
\code{"No return value"}.}

\item{ignore.overview}{whether to ignore lines containing the string
\code{"gives an overview of the package"}.}

\item{simplify}{whether to omit empty string vectors in the output.}
}
\value{
List containing lines of code that are different between the R file in the
two repositories.
}
\description{
Compare the source code of a given R file line by line, between the \pkg{TAF}
and \pkg{icesTAF} repositories.
}
\examples{
\dontrun{
tools.prod <- "~/git/ices/tools-prod"
diff_file("draft.data.R", tools.prod)
}

}
