% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_namespace.R
\name{diff_namespace}
\alias{diff_namespace}
\title{Compare Namespace}
\usage{
diff_namespace(tools.prod = "~/git/ices/tools-prod")
}
\arguments{
\item{tools.prod}{location of \code{TAF} and \code{icesTAF} repositories.}
}
\value{
List of \code{NAMESPACE} entries that are different between the repositories.
}
\description{
Compare entries in the \code{NAMESPACE} file of \pkg{TAF} and \pkg{icesTAF}
repositories.
}
\examples{
\dontrun{
tools.prod <- "~/git/ices/tools-prod"
diff_namespace(tools.prod)
}

}
