\name{read.columns}
\alias{read.columns}
\title{Read fixed width columns from a text file.}
\description{
	Reads specific columns from a fixed with text file.
}
\synopsis{
	read.columns (filename, columns)
}
\arguments{
	\item{filename}{The name of a file that is a fixed width data set. This file will be read using \code{read.fwf()}}
	\item{columns}{The columns (numbers) to read from the file.}
}
