% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.environment.R
\name{compare.environment}
\alias{compare.environment}
\title{Compare Environment}
\usage{
compare.environment()
}
\value{
List of objects that are different between the environments.
}
\description{
Compare objects in the environments \code{package:TAF} and
\code{package:icesTAF} in an R session.
}
\examples{
compare.environment()

}
